♻️ refactor(app): split router shell and lifecycle setup
Move the root shell, hash router setup, and theme/language lifecycle effects into dedicated app modules.

This keeps App.tsx focused on composition while preserving existing routing and global UI behavior.
