+++ src/App.tsx
@@ -1,63 +1,13 @@
-import { useEffect } from 'react';
-import { Outlet, RouterProvider, createHashRouter } from 'react-router-dom';
-import { LoginPage } from '@/pages/LoginPage';
-import { AppBackground } from '@/components/common/AppBackground';
-import { NotificationContainer } from '@/components/common/NotificationContainer';
-import { ConfirmationModal } from '@/components/common/ConfirmationModal';
-import { MainLayout } from '@/components/layout/MainLayout';
-import { ProtectedRoute } from '@/router/ProtectedRoute';
-import { useLanguageStore, useThemeStore } from '@/stores';
-
-function RootShell() {
-  return (
-    <>
-      <AppBackground />
-      <div className="app-content">
-        <NotificationContainer />
-        <ConfirmationModal />
-        <Outlet />
-      </div>
-    </>
-  );
-}
-
-const router = createHashRouter([
-  {
-    element: <RootShell />,
-    children: [
-      { path: '/login', element: <LoginPage /> },
-      {
-        path: '/*',
-        element: (
-          <ProtectedRoute>
-            <MainLayout />
-          </ProtectedRoute>
-        ),
-      },
-    ],
-  },
-]);
+import { AppLifecycle } from '@/app/AppLifecycle';
+import { AppRouter } from '@/app/AppRouter';
 
 function App() {
-  const initializeTheme = useThemeStore((state) => state.initializeTheme);
-  const language = useLanguageStore((state) => state.language);
-  const setLanguage = useLanguageStore((state) => state.setLanguage);
-
-  useEffect(() => {
-    const cleanupTheme = initializeTheme();
-    return cleanupTheme;
-  }, [initializeTheme]);
-
-  useEffect(() => {
-    setLanguage(language);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []); // 仅用于首屏同步 i18n 语言
-
-  useEffect(() => {
-    document.documentElement.lang = language;
-  }, [language]);
-
-  return <RouterProvider router={router} />;
+  return (
+    <>
+      <AppLifecycle />
+      <AppRouter />
+    </>
+  );
 }
 
 export default App;
