+++ src/app/AppLifecycle.tsx
@@ -1,0 +1,24 @@
+import { useEffect } from 'react';
+import { useLanguageStore, useThemeStore } from '@/stores';
+
+export function AppLifecycle() {
+  const initializeTheme = useThemeStore((state) => state.initializeTheme);
+  const language = useLanguageStore((state) => state.language);
+  const setLanguage = useLanguageStore((state) => state.setLanguage);
+
+  useEffect(() => {
+    const cleanupTheme = initializeTheme();
+    return cleanupTheme;
+  }, [initializeTheme]);
+
+  useEffect(() => {
+    setLanguage(language);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []); // 仅用于首屏同步 i18n 语言
+
+  useEffect(() => {
+    document.documentElement.lang = language;
+  }, [language]);
+
+  return null;
+}
