+++ src/app/AppRouter.tsx
@@ -1,0 +1,26 @@
+import { RouterProvider, createHashRouter } from 'react-router-dom';
+import { MainLayout } from '@/components/layout/MainLayout';
+import { LoginPage } from '@/pages/LoginPage';
+import { ProtectedRoute } from '@/router/ProtectedRoute';
+import { RootShell } from './RootShell';
+
+const router = createHashRouter([
+  {
+    element: <RootShell />,
+    children: [
+      { path: '/login', element: <LoginPage /> },
+      {
+        path: '/*',
+        element: (
+          <ProtectedRoute>
+            <MainLayout />
+          </ProtectedRoute>
+        ),
+      },
+    ],
+  },
+]);
+
+export function AppRouter() {
+  return <RouterProvider router={router} />;
+}
